requirements.txt
confluent-kafka==2.6.0
pymongo==4.8.0
pydantic==2.9.2
python-dotenv==1.0.1
colorama==0.4.6
rich==13.9.2